+++ BitmovinAnalyticsCollector/Classes/AVPlayer/Classes/AVPlayerAdapter.swift
@@ -5,8 +5,8 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
     static let timeJumpedDuplicateTolerance = 1_000
     static let maxSeekOperation = 10_000
     private static var playerKVOContext = 0
-    let drmPerformanceInfo: DrmPerformanceInfo?
     private let config: BitmovinAnalyticsConfig
+    private var drmPerformanceInfo: DrmPerformanceInfo?
     private var lastBitrate: Double = 0
     @objc private var player: AVPlayer
     let lockQueue = DispatchQueue.init(label: "com.bitmovin.analytics.avplayeradapter")
@@ -59,6 +59,14 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         resetState()
     }
 
+    private func updateDrmPerformanceInfo(_ playerItem: AVPlayerItem) {
+        if playerItem.asset.hasProtectedContent {
+            self.drmPerformanceInfo = DrmPerformanceInfo(drmType: DrmType.fairplay)
+        } else {
+            self.drmPerformanceInfo = nil
+        }
+    }
+
     private func startMonitoringPlayerItem(playerItem: AVPlayerItem) {
         statusObserver = playerItem.observe(\.status) {[weak self] (item, _) in
             self?.playerItemStatusObserver(playerItem: item)
@@ -67,6 +75,7 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         NotificationCenter.default.addObserver(self, selector: #selector(timeJumped(notification:)), name: NSNotification.Name.AVPlayerItemTimeJumped, object: playerItem)
         NotificationCenter.default.addObserver(self, selector: #selector(playbackStalled(notification:)), name: NSNotification.Name.AVPlayerItemPlaybackStalled, object: playerItem)
         NotificationCenter.default.addObserver(self, selector: #selector(failedToPlayToEndTime(notification:)), name: NSNotification.Name.AVPlayerItemFailedToPlayToEndTime, object: playerItem)
+        updateDrmPerformanceInfo(playerItem)
     }
 
     private func stopMonitoringPlayerItem(playerItem: AVPlayerItem) {
@@ -221,6 +230,11 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
 
         // isCasting
         eventData.isCasting = player.isExternalPlaybackActive
+
+        // DRM Type
+        if let drmType = self.drmPerformanceInfo?.drmType {
+            eventData.drmType = drmType
+        }
 
         // isLive
         let duration = player.currentItem?.duration